added back old features
- gdtot
- appdrive
- bot pm
- mirror log
- leech log
- torrent/direst/leech/mega limits
- telegraph/html
- source link from helios
- auto del links/upload msg
- 4gb bot pm
- force bot pm from helios
- split stats
- fsub
- title name
- sync with mltb

+++ bot/helper/telegram_helper/message_utils.py
@@ -7,7 +7,6 @@ from os import remove
 from bot import *
 from bot.helper.ext_utils.bot_utils import get_readable_message, setInterval
 
-
 def sendMessage(text: str, bot, message: Message):
     try:
         return bot.sendMessage(message.chat_id,
@@ -104,6 +103,18 @@ def auto_delete_message(bot, cmd_message: Message, bot_message: Message):
         sleep(AUTO_DELETE_MESSAGE_DURATION)
         try:
             # Skip if None is passed meaning we don't want to delete bot xor cmd message
+            deleteMessage(bot, cmd_message)
+            deleteMessage(bot, bot_message)
+        except AttributeError:
+            pass
+
+def auto_delete_upload_message(bot, cmd_message: Message, bot_message: Message):
+    if cmd_message.chat.type == 'private':
+        pass
+    elif AUTO_DELETE_UPLOAD_MESSAGE_DURATION != -1:
+        sleep(AUTO_DELETE_UPLOAD_MESSAGE_DURATION)
+        try:
+            # Skip if None is passed meaning we don't want to delete bot or cmd message
             deleteMessage(bot, cmd_message)
             deleteMessage(bot, bot_message)
         except AttributeError:
